SearchField cleanup and Fixes
