+++ styleguide/lib/styleguide.dart
@@ -4,4 +4,5 @@ export 'src/app_theme_data.dart';
 export 'src/components/app_card.dart';
 export 'src/components/loading_button.dart';
 export 'src/components/show_error_dialog.dart';
+export 'src/components/search_field_with_suggestions/search_field_with_suggestions.dart';
 export 'src/dimensions.dart';
